+++ ui/main/game/galactic_war/cards/gwaio_upgrade_subcommander_duplication.js
@@ -1,0 +1,28 @@
+define({
+  visible: _.constant(true),
+  describe: _.constant(
+    "!LOC:Sub Commander Duplication Tech adds an extra Commander to every Sub Commanders' army."
+  ),
+  summarize: _.constant("!LOC:Sub Commander Duplication Tech"),
+  icon: _.constant(
+    "coui://ui/main/game/galactic_war/shared/img/red-commander.png"
+  ),
+  audio: function () {
+    return {
+      found: "/VO/Computer/gw/board_tech_available_subcommander",
+    };
+  },
+  getContext: function (galaxy) {
+    return {
+      totalSize: galaxy.stars().length,
+    };
+  },
+  deal: function (system, context, inventory) {
+    var chance = 0;
+    if (inventory.hasCard("gwc_minion")) chance = 30;
+
+    return { chance: chance };
+  },
+  buff: function () {},
+  dull: function () {},
+});